Cast the sqrt to integer

+++ 0x1E-search_algorithms/100-jump.c
@@ -12,7 +12,7 @@
  */
 int jump_search(int *array, size_t size, int value)
 {
-	size_t step = sqrt(size), l = 0, r = step;
+	size_t step = (int)sqrt(size), l = 0, r = step;
 
 	if (array == NULL)
 		return (-1);
